+++ R/alterQset.R
@@ -351,7 +351,6 @@ renameQsetNames <- function(qseaSet, pattern, replacement = "") {
 #' # e.g., "Colon1_T","Colon1_N" -> "Colon1"
 #' poolSamples(exampleTumourNormal, mergeString = "_[TN]$")
 #' }
-#'
 #' @export
 poolSamples <- function(qseaSet, mergeString){