+++ tests/testthat/test-exampleQset.R
@@ -165,7 +165,25 @@ test_that("Analysing DMRs", {
                  pivotDMRsLonger() %>%
                  annotateWindows(genome = "hg38") %>%
                  summariseDMRsByGene() %>%
-                 dim(),c(21,4))
+                 ncol(),4)
+
+  expect_no_error(DMRs <- exampleTumourNormal %>%
+                    filter(type %in% c("LUAD","LUSC","CRC")) %>%
+                    calculateDMRs(variable = "type",
+                                  contrasts = "all"))
+  
+  expect_equal(DMRs %>% sliceDMRs(n = 1) %>% nrow(), 6)
+  expect_equal(DMRs %>% sliceDMRs(n = 2) %>% nrow(), 12)
+  
+  DMRs_log2 <- DMRs %>% sliceDMRs(n = 1, metric = log2FC)
+  DMRs_adjPval <- DMRs %>% sliceDMRs(n = 1, metric = adjPval)
+  DMRs_deltaBeta <- DMRs %>% sliceDMRs(n = 1, metric = deltaBeta)
+  DMRs_position <- DMRs %>% sliceDMRs(n = 1, metric = position)
+  
+  expect_false(all(DMRs_log2$start == DMRs_adjPval$start))
+  expect_false(all(DMRs_log2$start == DMRs_deltaBeta$start))
+  expect_false(all(DMRs_adjPval$start == DMRs_deltaBeta$start))
+  expect_false(all(DMRs_adjPval$start == DMRs_position$start))
 
   expect_no_error(DMRs %>% 
                     writeDMRsToBed(tempdir()))